bf16 support for per_channel bwd
[ghstack-poisoned]

+++ test/quantization/core/test_workflow_ops.py
@@ -110,10 +110,10 @@ def _fake_quantize_learnable_per_channel_affine_grad_reference(
     per_channel_zero_point = ((per_channel_zero_point.detach() + 0.5).clamp(quant_min, quant_max)).type(torch.int32)
     grad_X = _fake_quantize_per_channel_affine_grad_reference(
         dY, X, per_channel_scale, per_channel_zero_point, axis, quant_min, quant_max).to(device)
-    per_channel_scale = per_channel_scale.detach().type(torch.float)
+    per_channel_scale = per_channel_scale.detach().to(dtype)
 
-    grad_scale = torch.zeros([per_channel_scale.size(0)]).to(device)
-    grad_zero_point = torch.zeros([per_channel_zero_point.size(0)]).to(device)
+    grad_scale = torch.zeros([per_channel_scale.size(0)]).to(device, dtype=dtype)
+    grad_zero_point = torch.zeros([per_channel_zero_point.size(0)]).to(device, dtype=dtype)
 
     X_flattened = torch.unbind(X, dim=axis)
     dY_flattened = torch.unbind(dY, dim=axis)
@@ -127,14 +127,14 @@ def _fake_quantize_learnable_per_channel_affine_grad_reference(
         Xq_i = ((X_i / scale_i) + zero_point_i).round()
         Xfq_i = (Xq_i - zero_point_i) * scale_i
 
-        indicate_small_scale_i = (Xq_i < quant_min).float().to(device)
-        indicate_big_scale_i = (Xq_i > quant_max).float().to(device)
-        indicate_middle_scale_i = torch.ones(indicate_small_scale_i.shape).to(device) - \
+        indicate_small_scale_i = (Xq_i < quant_min).to(device, dtype=dtype)
+        indicate_big_scale_i = (Xq_i > quant_max).to(device, dtype=dtype)
+        indicate_middle_scale_i = torch.ones(indicate_small_scale_i.shape).to(device, dtype=dtype) - \
             indicate_small_scale_i - indicate_big_scale_i
 
-        indicate_saturate_zp_i = ((Xq_i < quant_min).float() +
-                                  (Xq_i > quant_max).float()).to(device)
-        indicate_unsaturate_zp_i = torch.ones(indicate_saturate_zp_i.shape).to(device) - \
+        indicate_saturate_zp_i = ((Xq_i < quant_min) +
+                                  (Xq_i > quant_max)).to(device, dtype=dtype)
+        indicate_unsaturate_zp_i = torch.ones(indicate_saturate_zp_i.shape).to(device, dtype=dtype) - \
             indicate_saturate_zp_i
 
         Xq_i = Xq_i.clamp(quant_min, quant_max)
